@@ -1,107 +1,70 @@
 source 'https://rubygems.org'
-ruby '2.3.1'
+git_source(:github) { |repo| "https://github.com/#{repo}.git" }
 
-# Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '4.2.10'
+ruby '2.7.2'
+
+# Bundle edge Rails instead: gem 'rails', github: 'rails/rails', branch: 'main'
+gem 'rails', '~> 6.1.3', '>= 6.1.3.1'
 # Use sqlite3 as the database for Active Record
-gem 'mysql2', '~> 0.4.0'
+gem 'sqlite3', '~> 1.4'
+# Use Puma as the app server
+gem 'puma', '~> 5.0'
 # Use SCSS for stylesheets
-gem 'sass-rails', '~> 5.0'
-# Use Uglifier as compressor for JavaScript assets
-gem 'uglifier', '>= 1.3.0'
-# Use CoffeeScript for .coffee assets and views
-gem 'coffee-rails', '~> 4.1.0'
-# See https://github.com/rails/execjs#readme for more supported runtimes
-# gem 'therubyracer', platforms: :ruby
-
-# Use jquery as the JavaScript library
-gem 'jquery-rails'
-# Turbolinks makes following links in your web application faster. Read more: https://github.com/rails/turbolinks
-gem 'turbolinks'
+gem 'sass-rails', '>= 6'
+# Transpile app-like JavaScript. Read more: https://github.com/rails/webpacker
+gem 'webpacker', '~> 5.0'
+# Turbolinks makes navigating your web application faster. Read more: https://github.com/turbolinks/turbolinks
+gem 'turbolinks', '~> 5'
 # Build JSON APIs with ease. Read more: https://github.com/rails/jbuilder
-gem 'jbuilder', '~> 2.0'
-# bundle exec rake doc:rails generates the API under doc/api.
-gem 'sdoc', '~> 0.4.0', group: :doc
-
-# Use ActiveModel has_secure_password
+gem 'jbuilder', '~> 2.7'
+# Use Redis adapter to run Action Cable in production
+# gem 'redis', '~> 4.0'
+# Use Active Model has_secure_password
 # gem 'bcrypt', '~> 3.1.7'
 
-# Use Unicorn as the app server
-# gem 'unicorn'
+# Use Active Storage variant
+# gem 'image_processing', '~> 1.2'
 
-# Use Capistrano for deployment
-# gem 'capistrano-rails', group: :development
+# Reduces boot times through caching; required in config/boot.rb
+gem 'bootsnap', '>= 1.4.4', require: false
 
 group :development, :test do
   # Call 'byebug' anywhere in the code to stop execution and get a debugger console
-  gem 'byebug'
-
-  gem 'bundler-audit', require: false
+  gem 'amazing_print'
+  gem 'byebug', platforms: %i[mri mingw x64_mingw]
 end
 
 group :development do
-  # Access an IRB console on exception pages or by using <%= console %> in views
-  gem 'web-console', '~> 2.0'
-
+  # Access an interactive console on exception pages or by calling 'console' anywhere in the code.
+  gem 'web-console', '>= 4.1.0'
+  # Display performance information such as SQL time and flame graphs for each request in your browser.
+  # Can be configured to work on production as well see: https://github.com/MiniProfiler/rack-mini-profiler/blob/master/README.md
+  gem 'listen', '~> 3.3'
+  gem 'rack-mini-profiler', '~> 2.0'
   # Spring speeds up development by keeping your application running in the background. Read more: https://github.com/rails/spring
+  gem 'annotate'
+  gem 'capistrano', require: false
+  gem 'foreman'
   gem 'spring'
-
-  gem 'rails-erd'
-
-  gem 'bullet'
-
-  gem 'lol_dba'
-
-  gem 'capistrano-rails'
-
-  gem 'capistrano-passenger'
-
-  gem 'capistrano-rvm'
-
-  gem 'capistrano-resque', "~> 0.2.2", require: false
-
-  gem 'better_errors'
-
-  # To fix: gem error
-  # gem 'wpdb_activerecord'
 end
 
-gem 'slim-rails', '3.1.1'
+group :test do
+  # Adds support for Capybara system testing and selenium driver
+  gem 'capybara', '>= 3.26'
+  gem 'selenium-webdriver'
+  # Easy installation and use of web drivers to run system tests with browsers
+  gem 'webdrivers'
+end
 
-gem 'awesome_print'
+# Windows does not include zoneinfo files, so bundle the tzinfo-data gem
+gem 'tzinfo-data', platforms: %i[mingw mswin x64_mingw jruby]
 
-gem 'simple_form'
-
-gem 'kaminari'
-
-gem 'handlebars_assets'
-
-gem 'settingslogic'
-
-gem 'resque'
-
-gem 'devise'
-
-gem 'ckeditor', '~> 4.2.4'
-
-gem 'dragonfly'
-
-gem 'chinese_pinyin'
-
-gem 'meta-tags'
-
+# Readonly ActiveRecord-esque
 gem 'active_hash'
 
-gem 'nokogiri'
-
-gem 'ransack'
-
-gem 'cancancan', '~> 1.10'
-
-gem 'rollbar'
-
-gem 'sitemap_generator'
-
-gem 'whenever'
-
+gem 'config'
+gem 'devise'
 gem 'rack-attack'
+gem 'resque'
+gem 'simple_form'
+gem 'sitemap_generator'
