@@ -1,8 +1,6 @@
 class Mailer < ApplicationMailer
-
   def contact(params)
     @params = params
-    mail to: Setting.email.info, subject: "[#{Setting.name} Contact] #{@params['subject']}"
+    mail to: Settings.email.info, subject: "[#{Settings.name} Contact] #{@params['subject']}"
   end
-
 end
