@@ -12,7 +12,7 @@ Devise.setup do |config|
   # Configure the e-mail address which will be shown in Devise::Mailer,
   # note that it will be overwritten if you use your own mailer class
   # with default "from" parameter.
-  config.mailer_sender = Setting.email.noreply
+  config.mailer_sender = Settings.email.noreply
 
   # Configure the class responsible to send e-mails.
   # config.mailer = 'Devise::Mailer'
